requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
scikit-learn>=1.4.0
Pillow>=10.2.0
numpy>=1.24.0
matplotlib>=3.8.0
